Fix weird class construction of MapToplevel
